Make sure screen locks on lid close.
Also, some notes on how to suspend the computer on lid close,
with a script option to do it when I need it.  (But suspend
must run as root, so ratpoison can't itself do the suspend.)

At ratpoison startup, launch the screen lock daemon as well as
a terminal and an emacs session.