Capture (highly reduced) screen shots each minute
Hugo Mougard noted in conversation on 2016-10-05 that perhaps window
title isn't sufficiently informative and that the window content
itself would be interesting to look at.  This presents a couple of
interesting challenges.  First, I'd like to keep the data available so
that I can train new classifiers as I learn more.  So I need to store
the windows I snap.  Second, I might be looking at something
confidential, and just setting file mode to 400 isn't sufficiently
reassuring.  So I reduce the image size to 100x100, which (visually)
seems a reasonable trade-off between image size and privacy.

To a human, the text is (generally) unreadable.  It is worth
remembering that some deep nets have managed to read such text anyway.