Note X cursor size, but in comment
I've had XCursor*size set to 96 on my laptop (starshine) for some time
without having committed to git.  I don't remember if it matters or
not.  So here I'm committing the line I've been using but leaving it
in a comment.  If it turns out that this is important, I should change
the dotfiles X11 install script to set it for HiDPI hosts and not for
LowDPI hosts.  (I should do something like that for dunst's rc as
well.

If, after my next X restart, I don't see a difference, I should remove
the line.