Improve lock and suspend functions
For reasons I don't fully understand, my laptop doesn't generally lock
or suspend on lid close.  And other times I believe it tries to
hibernate, from which it rarely recovers.

Improve the lock-suspend script to suspend correctly.  And add a new
i3 key binding, [C]-[A]-P, to trigger pm-suspend, with [C]-[A]-L only
locking.  In addition, have the lid monitor invoke i3-lock-suspend so
that the behaviour always happens there, reducing from two to one the
number of points I need to maintain for this behaviour.