Fix suspend condition
Sync after sleep, before suspend, just in case we don't wake.
Not that this has ever caused a problem.

Sometimes on wake we haven't quite slept, and there's a brief moment
when the top window takes input before screen lock.  The sync migth
just sort out that race condition, though that's a hack to be sure.