Use kitty instead of sakura
I picked sakura when I started using sway because the default choice
in sway was foot, which bothered me for reasons I don't remember.

Today I discovered kitty, which seems to be quite good.  It has a bit
of a learning curve, on a par with urxvt, so the choice is tentative.
But it already seems better than sakura.