Make my desktop suspendable under sway
It was a bit tricky, as most of the time I don't want it to suspend
during the day, as I'm doing stuff and want to come back to progress.
But when leaving for a while, I want to give myself the opportunity to
indicate that I'd like to suspend.